Save "aspect ratio" information for images so that pages can attempt to handle variable image shapes
Eg Suppose you have several images for a product of variying shape
but need the image's bounding box to stay at one size. Paperclip
will give you images of a known width for a given image format
(if you set the :styles for 'product', 'thumb', ...  appropriately),
but height can differ. This fix saves the height/width ratio as a
percentage in the (unused) attachment_size field , so you can
obtain an estimate of the height from the width.