600 - create backup files when copying during spree:upgrade
Two changes here:

"mirror_files" has been generalised to work on single files (for consistency),
and will create backups with a ~ suffix if passed a flag. It also only copies
new files when the content has changed (before, it was testing identity up to
links)

in the upgrade task, the same operation is used on all things to be copied
(consistency...), and the copying operation will create backups instead of
silent over-writing.

Note that FileUtils.cp etc have certain options hard-coded, and none of
these is the -b option, so some DIY is needed.

[#600 state:patched]